# test: simplify and stabilize e2e tests