--- a/jq-build-bin
+++ b/jq-build-bin
@@ -1,0 +1,1 @@
+/nix/store/7jvwwz45qap70i6asxzcnbshhz88a5sv-jq-1.8.1-bin