--- a/jq-build-man
+++ b/jq-build-man
@@ -1,0 +1,1 @@
+/nix/store/laqc7c2c9j09vm1w58vmj633vc2wb3wv-jq-1.8.1-man