… .github/actions/setup-install/action.yml
… .github/workflows/release.yml
… .github/workflows/run-tests.yml
… README.md
… docs/nix.md
… package.json
… packages/app/components.json
… packages/app/index.html
… packages/app/oxlint.json
… packages/app/package.json
… packages/app/playwright.config.ts
… packages/app/postcss.config.js
… packages/app/public/favicon.svg
… packages/app/public/jq
… packages/app/public/jq-version.json
… packages/app/public/jq.js
… packages/app/public/jq.wasm
… packages/app/public/jq_1-8-1.js
… packages/app/public/jq_1-8-1.wasm
… packages/app/pwa-assets.config.ts
… packages/app/src/PWABadge.tsx
… packages/app/src/assets/hason-screenshot.png
… packages/app/src/assets/react.svg
… packages/app/src/components/help-button.tsx
… packages/app/src/components/help-drawer.tsx
… packages/app/src/components/share-button.tsx
… packages/app/src/components/theme-provider.tsx
… packages/app/src/components/theme-switcher.tsx
… packages/app/src/components/theme-toggle.tsx
… packages/app/src/components/ui/badge.tsx